requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.8.0
astroid==3.2.4
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
branca==0.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
dnspython==2.6.1
fastapi==0.115.8
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
folium==0.19.4
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html2text==2024.2.26
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.2.2
packaging==24.1
pandas==2.2.3
pkce==1.0.3
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycurl==7.45.3
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pylint==3.2.6
pymongo==4.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.45.3
tomlkit==0.13.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.16
uvicorn==0.34.0
Werkzeug==3.0.3
wikipedia==1.4.0
Wikipedia-API==0.7.1
wptools==0.4.17
xyzservices==2025.1.0
